devicepool_config_generator/pyproject.toml
[tool.poetry]
name = "devicepool_config_generator"
version = "0.1.0"
description = ""
authors = ["Andrew Erickson <aerickson@mozilla.com>"]

[tool.poetry.dependencies]
python = "^3.5"  # Compatible python versions must be declared here
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
